requirements.txt
asgiref>=3.3.1
dj-database-url==0.5.0
Django~=4.1
gunicorn==20.0.4
Pillow>=9
pytz==2020.1
sorl-thumbnail==12.8.0
sqlparse==0.3.0
whitenoise==5.2.0
pyyaml
uritemplate

psycopg2-binary==2.9.5
django-debug-toolbar
requests~=2.28.1
# python-dotenv
environs~=9.5.0
djangorestframework
django-cors-headers
django-filter

social-auth-app-django
# for requests stats
django-request
djangorestframework-simplejwt